remote_kitchen/requirements.txt
asgiref==3.7.2
braintree==4.26.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
fontawesomefree==6.5.1
idna==3.6
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
